+++ ui/v2.5/src/components/FrontPage/FrontPage.tsx
